Separate embedded server from client config
Make the embedded server support separate from the client connection.
The idea is to allow a server to be created, without necessarily using
the InVMConnectorFactory or to use a InVMConnectorFactory with a server
configured outside of the autoconfiguration.

Also added an enabled property so allow the embedded server to be
disabled entirely.