Move ConnectionFactory creation
Move ConnectionFactory creation logic to the top configuration and
deduce the factory to use when no mode is not explicitly set. Also
refactored properties class so that it can be used when jms-server.jar
is not present.

Improved exception message if embedded mode is specified but the
InVMConnectorFactory class cannot be found.

(several tests left to fixup)