Add active profiles argument
This commit adds an additional property to the run goal of the maven
plugin that allows to specify the profile(s) to activate. Those can be set
on the command line in a convenient way. For instance, this would enable
the prod and foo profiles:

mvn spring-boot:run -Drun.profiles=prod,foo

Relates to gh-1176