Add group discriminant in case of conflict
Prior to this commit, the repackage goal silently ignore the case of
two libraries having the same name and version but a different group.
As a result, the second library was overwriting the first one in the
repackaged jar.

This commit updates Library to include the notion of "group" which is
filled by both the Maven and Gradle plugins. In case of conflict that
value is used to update the name of the library in the repackaged
archive.

Fixes gh-1475