Fix order of JMS auto configuration
There are multiple sources for a JMS ConnectionFactory:

1. JNDI
2. HornetQ, either from an embedded broker that is started with the
application or connecting to an existing broker
3. ActiveMQ (same options as HornetQ)

The last two auto configurations must run after the JTA auto
configuration has completed as this may register additional beans that
are necessary to enable XA.

Previously, the HornetQ embedded broker would start regardless of the
presence of a ConnectionFactory as a ConditionalOnMissingBean was
missing. Furthermore, there was no order condition for the JNDI auto
config so it may just run after one of the broker has been found.

JNDI takes now precedence to be consistent with the regular DataSource
auto configuration.

Fixes gh-1821