Add start/stop goals to the maven plugin
This is an attempt at finalizing the feature defined in gh-2525.

Including fork support brought a lot of complexity for no good reason
and right now the Maven process is unable to connect to the MBean server
of a different process for no good reason. jconsole can connect just fine
and the same code works "in-process".