Cache auto-configuration
Initial support for cache auto-configuration that is triggered when the
`EnableCaching` annotation is added to a Spring Boot application.

If no explicit configuration is provided, the environment is checked for
the best suited cache implementation, that is:

* Redis if a `RedisTemplate` is defined in the context
* Generic if at least one `Cache` is defined in the  context
* Simple as a fallback option, using concurrent maps

It is also possible to set the cache provider to use in case of conflict.

Fixes gh-2633