Expose caching-specific infrastructure
Expose the underlying CacheManager if Boot auto-configures it. This is
the case for ehCache, hazelcast and JCache. This change has two side
effects:

1. It is now possible to customize the underlying cache infrastructure
and let Boot only wrap it in the Spring's CacheManager abstraction. No
customizations are applied if the caching-specific service is customized
2. Such infrastructure is disposed when the application terminates as
it is now defined as `@Bean` and both `close()` and `shutdown()` methods
are invoked if present on the target type.

2 can be troublesome as there is no way to control or customize that if
we ever wanted to do it in the future.

Closes gh-2848