DevTools safe mode
DevTools check the current thread to figure out whether it has to enable
the reload feature. When the Maven plugin runs a Spring Boot application
without forking, the classloader is not the one that DevTools expect and
as a result, the feature is not enabled.

A new system property `DEVTOOLS_SAFE_MODE` can be set prior of starting
the application to disable all checks. The Maven plugin now uses that
property accordingly.

Closes gh-3315