Remove HornetQ support
Spring Boot build to use against the sample app demonstrated in the
Master Spring Boot auto-configuration talk. In theory, this is not
strictly necessary as the sample app could just add an exclude to the
auto-configuration as follows:

@SpringBootApplication(exclude=HornetQAutoConfiguration.class)

That being said, removing completely the support has the advantage of
only showing the demo auto-config in the logs as well as the keys
suggestion in the IDE.

A build will not pass until the tests are skipped since other components
of Boot obviously rely on the support that was just removed. To build
that locally as a companion of the code of the talk:

mvn clean install -DskipTests