Collect and display build information
This commit updates the build plugin to generate a
`META-INF/boot/build.properties` file with various build-specific
settings (group, artifact, version and build time).

A new `BuildInfo` bean is automatically exposed when such a file is
present.

If that information is available, an `InfoContributor` is automatically
created to expose that information under the `build` key

Closes gh-2559