Initiate GitProperties
Migrate GitInfo to GitProperties. There are two issues with the current
proposal:

1. Dates are no longer handled by GitProperties because it only knows
about "String". We could clean the String value (parsing the date and
always storing the epoch time), but it wouldn't help much in the info
contributor as there is no way for the binder to know that this
particular key is a Date. Having a `InfoProperties` that work with an
underlying Map<String,Object> would allow that use case
2. InfoProperties need to be translated to a source for the binder in
the `InfoContributor` implementation. The easiest is to have a
`toProperties` method that exposes a `java.util.Properties`

This commit also updates the git info contributor to be customized via
 `management.info.git.mode` which can be `simple` or `full`.