Fix build with Hibernate 5
This commit updates our test suite to be able to run on both Hibernate 4
and Hibernate 5. Two important changes made this change necessary:

* Spring Boot no longer provides a naming strategy for Hibernate 5 so it
defaults to the one provided by Hibernate. The main difference is the
physical name that is provided for camel case properties
* If the `@GeneratedValue` uses the `AUTO` strategy, Hibernate 5 uses
sequence by default. This commit forces the strategy to `IDENTITY` in
order to be compliant with what Hibernate 4 does by default

See gh-2763