Register error pages by default
This commit improves the developer experience by scanning and registering
error pages automatically. If a template with the name of a status code
is available from the `error` package, it is used as a view when such
error status needs to be handled by the error controller.

If an html static asset is located from the `error` package with the same
naming semantics, it is registered as an error page. Templates take
precedence over static assets: if a given status is handled by both a
template and a static asset, no error page is registered.

For instance, adding a `src/main/resources/static/error/404.html` file is
all that is required to display it if the server returns a 404.

Catch-all templates can also be defined so a `4xx` or `5xx` template
would be used if no other corresponding template exists.

Closes gh-2691