Fix CNFE caused by verifier loading SpringValidatorAdapterWrapper
I think the CNFE was caused by the verifier. It's certainly caused
by the statement that returns a SpringValidatorAdapterWrapper. I
believe the verifier loads the class that's being returned to check
that its type is compatible with the method's return type. Loading
SpringValidatorAdapterWrapper loads SpringValidatorAdapter which
loads javax.validation.Validator.

Moving all of the logic into your existing Jsr303ValidatorHandler
fixes the problem.