Be more defensive before flagging the Validator Primary
This commits ensure that the auto-configured Validator will not be flagged
as `@Primary` if a Spring Validator bean is itself `@Primary`.

`LocalValidatorFactoryBean` exposes 3 contracts and we're only checking
for the absence of a `javax.validation.Validator` to auto-configure a
`LocalValidatorFactoryBean`. If no standard JSR validator exists but a
Spring's `Validator` exists and is primary, we shouldn't flag the
auto-configured one as `@Primary`.

Previous iterations on this feature have made sure that we'll
auto-configure at most one `javax.validation.Validator` so not flagging it
`@Primary` is no problem.

See gh-8495