Move technology specific endpoints to extensions
This commit renames WebEndpoint and JmxEndpoint to WebExtension and
JmxExtension respectively. As a result, an endpoint has to be declared
with Endpoint now, regardless of how it has to be exposed.

JmxEndpointDiscoverer and WebEndpointDiscoverer are now specialization
of the base EndpointDiscoverer with most of the merging logic shared
at that level.

Because an extension does not have the same endpoint attributes as
a regular endpoint, that information is no longer provided to an
operation factory. Rather, the id of the endpoint is given, which is
what the current implementations were using anyway.