Provider generic infrastructure for merge and validation
This commit moves the endpoint's validation and merge one level up so
that it can be shared consistently. Tech-specific discoverer have now
to provide a function that generates a unique key for the operation.

The base `EndpointDiscoverer` now provides a description of each
endpoint that provides the consolidated endpoint (if necessary) and any
error. This can be used by the actual implementation to throw a
meaningful error.