Expose Jolokia directly rather than via an endpoint
Jolokia is a 100% web concern and does not fit in the Endpoint
infrastructure. This commit removes `JolokiaMvcEndpoint` and exposes
the servlet directly instead.

This is a work in progress, security needs to be added and the
configuration could use some love (or at least some alignment with
what we do with the H2 console).

See gh-9843