Default Hibernate DDL auto to none with Flyway/Liquibase
This commit adds a strategy interface to customize the default DDL auto
value for a DataSource. The Hibernate auto-configuration uses it to set
it to "none" if a mechanism to initialize the DataSource is
found and "none" otherwise.

Both Flyway and Liquibase implements that strategy interface and
register it in the context accordingly.

Closes gh-9262