Allow to tune how ConfigurationProperties are bound
This commit exposes a new API that allows a sysem to adapt or replace the
component that binds `@ConfigurationProperties` annotated objects.

`ConfigurationPropertiesBinder` is now an interface for easier reuse and
wrapping that is created via a `ConfigurationPropertiesBinderFactory`. If
a bean with the name of the factory's class is exposed, the
post-processor will use it instead of the default.

The post-processor is also reclaiming space once the context has
refreshed by calling `destroy` on the factory with the binder it has
created. A factory implementation may decide to keep the instance if
further use of the binder is expected.

Closes gh-10685