Polish "Introduce HealthIndicatorRegistry"
This commit is work in progress and polishes the initial submission by
making sure that the CompositeHealthIndicator is also a registry.

The current infrastructure prevents the composite to be registered as a
bean as it would be an extra HealthIndicator in the context and
therefore may be taken into account as a regular indicator. The registry
on the other hand must be registered in the context so that users can
manipulate its content.

Ultimately, both those objects share common base features that we should
not duplicate. This commit makes sure that the registry has a `health()`
method but does not implement HealthIndicator. The composite now extends
from the registry and implements HealthIndicator, simply delegating to
that parent method.

It's unclear at this point if we want to keep this arrangement as the
duplication of factories is a bit annoying.

See gh-4965