Add auto-configuration support for TaskExecutor
This commit adds support for configuring a ThreadPoolTaskExecutor as
well as auto-configuration when it makes sense.

A ThreadPoolTaskExecutor can be easily built using TaskExecutorBuilder
that is available as a bean. The default builder can be initialized
using the `spring.task.*`namespace.

As the necessary classes to setup such pool are always present, only
the builder is made available by default. If `@EnableAsync` is detected
and no TaskExecutor was configured, the one produced by the builder is
exposed automatically.

See gh-1563