Add command to extract application in an optimized structure
This commit is a WIP that hacks the layertools jar to provide an
additional extract command that creates the following structure:

my-dir
├── application
│   └── my-app-1.0.0-SNAPSHOT.jar
├── dependencies
│   ├── ...
│   ├── spring-context-6.1.0-RC2.jar
│   ├── spring-context-support-6.1.0-RC2.jar
│   ├── ...
└── run-app.jar

With such a structure in place, the application can start with
`java -jar run-app.jar`.

Classes that were originally in `BOOT-INF/classes` are jared in the
`application` folder with the same name as the original repackaged jar.

Only dependencies that are listed in the `classpath.idx` file are
copied, and the `run-app.jar` contains a MANIFEST that list those libs
in the same order.