Add execution context in Quartz job detail
This commit improves the Quartz Actuator endpoint to specify whether a
particular job is running. For a job that's running, the related trigger
has some details such as the fire time, whether it's recovering and the
re-fire count.

To make things more consistent, the detail of a trigger has now an
executions that contains detail about the previous, current, and next
executions.

Closes gh-43226