Add metadata support for types in arbitrary modules
Previously, if a ConfigurationProperties had a nested type or was
extending from a type located outside the compilation unit, no
metadata discovered on the source code was available. This typically
happens when such a type reside in another module.

This commit introduces `@ConfigurationPropertiesSource` as a way to
annotate such type and have metadata generated for them in their own
module. Type-metadata is generated as one file per type and is reused
transparently whenever that type is used.

As for module metadata, an additional file can be crafted manually and
will be merged when the metadata for the type is generated.