Add support for Spring Batch MongoDB
This commit adds support for running Spring Batch job with a MongoDB
store. It aligns as much as possible to the JDBC counterpart, with a
new spring-boot-starter-data-mongodb, and its test-based equivalent.

As we do not have a way to initialize a MongoDB store at the moment,
this commit adds a conservative approach of executing commands defined
by the standard Spring Batch schema script.

Closes gh-43226